Added new listener and finished TestServiceModel
The new listener is used to detect any problem with database connection.
There is also an exception for when there is a problem with the online
service from the ServiceModel class.
Finally, the tests for the ServiceModel class should be finished,
although there could be some additions in the future